bb:Coverity fix for restart case with undefined contribid value
(cherry picked from commit d9b811a4fc3d199f43e909a73f2f40b0b4322a28)
(cherry picked from commit 20b9c6bd0f961d4d876d1a76ebe849d250527387)

+++ bb/src/BBLV_Info.cc
@@ -326,7 +326,7 @@ void BBLV_Info::removeFromInFlight(const string& pConnectionName, const LVKey* p
             //       no extents for this file currently in-flight...  If so, delay for a bit...
             uint32_t i = 0;
             int l_DumpOption = DO_NOT_DUMP_QUEUES_ON_VALUE;
-            while (extentInfo.moreExtentsToTransferForFile(pExtentInfo.getHandle(), pExtentInfo.getContrib(), pExtentInfo.getSourceIndex(), THIS_EXTENT_IS_IN_THE_INFLIGHT_QUEUE, l_DumpOption))
+            while (extentInfo.moreExtentsToTransferForFile((int64_t)pExtentInfo.getHandle(), (int32_t)pExtentInfo.getContrib(), pExtentInfo.getSourceIndex(), THIS_EXTENT_IS_IN_THE_INFLIGHT_QUEUE, l_DumpOption))
             {
                 unlockTransferQueue(pLVKey, "removeFromInFlight - Waiting for inflight queue to clear");
                 {
